파일 분리, 테스트 통과
